Formatted Code

+++ contracts/orderbook/Keepers/OrderBookInterface.sol
@@ -1,28 +1,42 @@
 pragma solidity ^0.8.4;
 
-interface IOrderBook{
-	function getOrders(uint start, uint count) external view returns(OpenOrder[] memory);
-	function prelimCheck(address trader, uint orderID) external view returns(bool);
-	function executeOrder(address payable keeper, address trader, uint orderID) external;
-	function getTotalActiveOrders() external view returns(uint);
-    struct OpenOrder{
+interface IOrderBook {
+    function getOrders(uint256 start, uint256 count)
+        external
+        view
+        returns (OpenOrder[] memory);
+
+    function prelimCheck(address trader, uint256 orderID)
+        external
+        view
+        returns (bool);
+
+    function executeOrder(
+        address payable keeper,
+        address trader,
+        uint256 orderID
+    ) external;
+
+    function getTotalActiveOrders() external view returns (uint256);
+
+    struct OpenOrder {
         address trader;
         bytes32 loanID;
         address iToken;
-		address loanTokenAddress;
-        uint price;
-        uint leverage;
-        uint loanTokenAmount;
-        uint collateralTokenAmount;
+        address loanTokenAddress;
+        uint256 price;
+        uint256 leverage;
+        uint256 loanTokenAmount;
+        uint256 collateralTokenAmount;
         bool isActive;
         address base;
-        uint orderType;
+        uint256 orderType;
         bool isCollateral;
-        uint orderID;
+        uint256 orderID;
         bytes loanData;
     }
-	struct OrderQueue{
-		address trader;
-		uint orderID;
-	}
+    struct OrderQueue {
+        address trader;
+        uint256 orderID;
+    }
 }
